Disable decay functions for older worker versions
